build!: set vsce as a peer dependency (#67)
This is needed so the users can choose which vsce version to use. This is a breaking change because users who did not install vsce as a [dev]dependency and [are not using npm@7](https://github.com/npm/rfcs/blob/latest/implemented/0025-install-peer-deps.md) will have to install vsce manually.